# endpoint/id: use strings.IndexByte