Use `strings.IndexByte` instead of `strings.Index` to look for a single byte. This improves `BenchmarkSplitID` as follows:

```
name       old time/op    new time/op    delta
SplitID-8    62.5ns ±61%    41.1ns ± 5%  -34.14%  (p=0.000 n=10+9)
```